backend/requirements.txt
Flask
Flask-CORS
torch
torchvision
Pillow
